"About" page refactored
I decided to introduce the site.* variables to simplify the procedure
of writing technical information. Look at the end of site configuration
file for informational variables.

This will allow to correct technical information in all places at once
in case it will be needed and also allow to avoid mistakes in text.